requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
transformers==4.34.0
torch==2.0.1
Pillow==10.0.1
speechrecognition==3.8.1
pyttsx3==2.90
ultralytics==8.0.157
Flask-Cors==3.0.10
eventlet
